+++ docs/CHANGELOG.md
@@ -3,6 +3,8 @@
 # Next
 
 * Updated dependency to **gopjrt** 0.4.5
+* Package `downloader`:
+  * Added support for user-agent.
 
 # v0.15.0 - 2024/11/01 Some API clean up; Added support for ONNX model conversion.